refactor(config): remove math-preprocess compatibility layer
Move markdownCompatibilityPreprocess logic into config/mdsvex/index.js
as createMarkdownPreprocess(), exported via mdsvex.config.js.

- Delete config/mdsvex/compatibility-preprocess.js
- Delete src/lib/math-preprocessor.js and its test
- Move test to tests/markdown/math-preprocess.test.ts
- Update svelte.config.js to use markdownPreprocess
- Update package.json test:markdown script
- Update vitest include pattern for tests/markdown
- Update AGENTS.md and CLAUDE.md references
