+++ mdsvex.config.js
@@ -1,6 +1,7 @@
 import { createMarkdownConfig } from './config/mdsvex/index.js';
 
-const { config, ctx } = await createMarkdownConfig();
+const { config, ctx, preprocess } = await createMarkdownConfig();
 
 export const markdownCtx = ctx;
+export const markdownPreprocess = preprocess;
 export default config;
