--- src/lib/math-preprocessor.js
@@ -1,1 +1,0 @@
-export { markdownCompatibilityPreprocess as mathPreprocess } from '../../config/mdsvex/compatibility-preprocess.js';
